Gas works now

+++ app/src/main/java/com/github/heronerin/secureroute/tabs/addPages/AddTripFragment.java
@@ -31,7 +31,7 @@ public class AddTripFragment extends AbstractAddPage {
         String odometerText = ((EditText)getActivity().findViewById(R.id.odometerV)).getText().toString();
         return !odometerText.isEmpty()
                 && (
-                revise != null || DataBase.instance.getLastWithOdometer() == null || Integer.valueOf( odometerText ) >= DataBase.instance.getLastWithOdometer().odometer
+                revise != null || DataBase.instance.getLastWithOdometer() == null || Long.valueOf( odometerText ) >= DataBase.instance.getLastWithOdometer().odometer
         );
     }
 
@@ -45,7 +45,7 @@ public class AddTripFragment extends AbstractAddPage {
                 isEnding ? lastKnownEnd.databaseId : -1,
                 ((EditText) getActivity().findViewById(R.id.noteField)).getText().toString(),
                 new JSONArray(),
-                Integer.valueOf(((EditText) getActivity().findViewById(R.id.odometerV)).getText().toString())
+                Long.valueOf(((EditText) getActivity().findViewById(R.id.odometerV)).getText().toString())
         );
     }
